Work in Progress: We're going to have a GameStarter after all.
The more I look at it the more it circles back to the current design.
